Move GetAllInterfaces to ReflectionUtils
