+++ src/Mapster/Utils/ReflectionUtils.cs
@@ -84,16 +84,55 @@ namespace Mapster
         {
             var bindingFlags = BindingFlags.Instance | accessorFlags;
 
-            var properties = type.GetProperties(bindingFlags)
+            IEnumerable<IMemberModelEx> getPropertiesFunc(Type t) => t.GetProperties(bindingFlags)
                 .Where(x => x.GetIndexParameters().Length == 0)
                 .Where(x => !requireSetter || x.CanWrite)
                 .Select(CreateModel);
 
-            var fields = type.GetFields(bindingFlags)
+            IEnumerable<IMemberModelEx> getFieldsFunc(Type t) => t.GetFields(bindingFlags)
                 .Where(x => !requireSetter || !x.IsInitOnly)
                 .Select(CreateModel);
 
+            IEnumerable<IMemberModelEx> properties;
+            IEnumerable<IMemberModelEx> fields;
+
+            if (type.IsInterface)
+            {
+                IEnumerable<Type> allInterfaces = GetAllInterfaces(type);
+                properties = allInterfaces.SelectMany(currentInterface => getPropertiesFunc(currentInterface));
+                fields = allInterfaces.SelectMany(currentInterface => getFieldsFunc(currentInterface));
+            }
+            else
+            {
+                properties = getPropertiesFunc(type);
+                fields = getFieldsFunc(type);
+            }
+
             return properties.Concat(fields);
+        }
+
+        // GetProperties(), GetFields(), GetMethods() do not return properties/methods from parent interfaces,
+        // so we need to process every one of them separately.
+        public static IEnumerable<Type> GetAllInterfaces(Type interfaceType)
+        {
+            var allInterfaces = new List<Type>();
+            var interfaceQueue = new Queue<Type>();
+            allInterfaces.Add(interfaceType);
+            interfaceQueue.Enqueue(interfaceType);
+            while (interfaceQueue.Count > 0)
+            {
+                var currentInterface = interfaceQueue.Dequeue();
+                foreach (var subInterface in currentInterface.GetInterfaces())
+                {
+                    if (allInterfaces.Contains(subInterface))
+                    {
+                        continue;
+                    }
+                    allInterfaces.Add(subInterface);
+                    interfaceQueue.Enqueue(subInterface);
+                }
+            }
+            return allInterfaces;
         }
 
         public static bool IsCollection(this Type type)
